requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
Pillow
Jinja2
python-multipart

